#22 Implement object tracking of cones
I used the findContours method to implement object tracking of cones based on their binary masks. Recall that the binary masks were obtained after HSV filtering. The method findContours uses edge detection, and therefore it is quite sensitive to the noise in the image. I had to tune the Gaussian blurring kernel size and standard deviation to achieve good results. It needs to be mentioned that I have used the contour mode RETR_TREE, which sorts the detected contours based on hierarchy. We are using the biggest contour with children which is indexed at 0.